mensajes de alerta

+++ src/app/cita/components/registro-citas/registro-citas.component.ts
@@ -84,12 +84,31 @@ export class RegistroCitasComponent implements OnInit {
   }
 
   async onGoToDelete(citaId: string): Promise<void> {
-    try {
-      await this.citasService.onDeleteCita(citaId);
-      alert('Deleted: Cita ID:' + citaId);
-    } catch (error) {
-      console.log(error);
-    }
+    // try {
+    //   await this.citasService.onDeleteCita(citaId);
+    //   alert('Deleted: Cita ID:' + citaId);
+    // } catch (error) {
+    //   console.log(error);
+    // }
+    
+    Swal.fire({
+      title: 'Estas seguro de borrar este paciente?',
+      text: "Esta accion no se puede revertir!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, borrar registro!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.citasService.onDeleteCita(citaId);
+        Swal.fire(
+          'Borrado!',
+          'El registro del paciente ha sido borrado exitosamente.',
+          'success'
+        )
+      }
+    })
   }
 
   cambiarPagina() {
